+++ include/mxnet/engine.h
@@ -41,26 +41,8 @@ class Engine;
 
 /*! \brief namespace of engine internal types. */
 namespace engine {
-/*! \brief base class of engine variables.*/
-struct Var {
-  virtual size_t version() {
-    return version_;
-  }
-  virtual ~Var() = default;
-  /*!
-   * \brief cast variable to derived type T
-   * \tparam T the type we want to cast into.
-   * \return A casted variable.
-   */
-  template <typename T>
-  inline T* Cast();
-  /*!
-   * \brief version number of the var. Every time the object it is associated with
-   * is modified, the version number is incremented by 1.
-   */
-  size_t version_{0};
-};  // struct Var
-
+/*! \brief Internal representation of variable. */
+struct Var;
 /*! \brief Internal representation of operator.  */
 struct Opr;
 /*! \brief Variable pointer type, usually hold by user used to specify dependencies. */
