+++ src/engine/engine_impl.h
@@ -33,6 +33,20 @@
 namespace mxnet {
 namespace engine {
 
+/*! \brief base class of engine variables, used for type checking */
+struct Var {
+#if ENGINE_DEBUG
+  virtual ~Var() = default;
+#endif  // ENGINE_DEBUG
+  /*!
+   * \brief cast variable to derived type T
+   * \tparam T the type we want to cast into.
+   * \return A casted variable.
+   */
+  template <typename T>
+  inline T* Cast();
+};  // struct Var
+
 /*! \brief base class of engine operators, used for type checking */
 struct Opr {
 #if ENGINE_DEBUG
