+++ src/engine/naive_engine.cc
@@ -28,23 +28,9 @@
 #include "./engine_impl.h"
 #include "../profiler/profiler.h"
 #include "./openmp.h"
-#include "../common/object_pool.h"
 
 namespace mxnet {
 namespace engine {
-
-/*!
- * \brief var used in Naive Engine for tracking the version
- * of the objects it is associated with.
- */
-class NaiveVar final
-    : public Var, public common::ObjectPoolAllocatable<NaiveVar> {
- public:
-  inline static NaiveVar* CastFromBase(Var* ptr) {
-    return ptr->Cast<NaiveVar>();
-  }
-};  // class NaiveVar
-
 
 // implement naive engine
 class NaiveEngine final : public Engine {
@@ -85,7 +71,8 @@ class NaiveEngine final : public Engine {
 
   // new variables
   VarHandle NewVariable() override {
-    return NaiveVar::New();
+    size_t v = ++counter_;
+    return reinterpret_cast<VarHandle>(v);
   }
 
   OprHandle NewOperator(AsyncFn fn,
@@ -159,10 +146,6 @@ class NaiveEngine final : public Engine {
       opr->opr_profile.reset(new profiler::ProfileOperator(opr->opr_name, attrs.release()));
       opr->opr_profile->start(exec_ctx.dev_type, exec_ctx.dev_id);
     }
-    // increment mutable var version
-    for (auto var : mutable_vars) {
-      ++var->version_;
-    }
     if (exec_ctx.dev_mask() == gpu::kDevMask) {
 #if MXNET_USE_CUDA
       size_t dev_id = static_cast<size_t>(exec_ctx.dev_id);
@@ -188,12 +171,8 @@ class NaiveEngine final : public Engine {
   }
 
   void DeleteVariable(SyncFn delete_fn, Context exec_ctx, VarHandle var) override {
-    NaiveVar* naive_var = NaiveVar::CastFromBase(var);
-    this->PushAsync([delete_fn, naive_var](RunContext ctx, CallbackOnComplete on_complete) mutable {
-        delete_fn(ctx);
-        NaiveVar::Delete(naive_var);
-        on_complete();
-      }, exec_ctx, {}, {var}, FnProperty::kDeleteVar, 0, "DeleteVariable");
+    this->PushSync(delete_fn, exec_ctx, {}, {var},
+                   FnProperty::kNormal, 0, "DeleteVariable");
   }
 
   void WaitForVar(VarHandle var) override {
@@ -213,6 +192,8 @@ class NaiveEngine final : public Engine {
   }
   // whether action is completed
   bool req_completed_;
+  // counter
+  std::atomic<size_t> counter_{0};
   /*! \brief whether it is during shutdown phase*/
   std::atomic<bool> shutdown_phase_{false};
   // CPU stream
