+++ src/engine/threaded_engine.cc
@@ -130,9 +130,6 @@ inline bool ThreadedVar::CompleteWriteDependency(Dispatcher dispatcher) {
     assert(pending_write_ != nullptr);
     CHECK_EQ(num_pending_reads_, kWriteTriggered);
 
-    // increment version number
-    ++version_;
-
     // really delete
     if (to_delete_) {
       VersionedVarBlock *head = pending_write_->next;
@@ -167,7 +164,7 @@ inline bool ThreadedVar::CompleteWriteDependency(Dispatcher dispatcher) {
   }
   // This is outside of lock scope
   // Be very carful, pending_write_ and num_pending_reads_
-  // can change now, do not rely on these two variables.
+  // can change now, do not reply ont the two variables.
   // The linked list \in [old_pending_write, end_of_read_chain)
   // is already detached from this Var.
   // So it is safe to modify these
@@ -197,11 +194,6 @@ inline void ThreadedVar::SetToDelete() {
 inline bool ThreadedVar::ready_to_read() {
   std::lock_guard<std::mutex> lock{mutex_};
   return this->is_ready_to_read();
-}
-
-inline size_t ThreadedVar::version() {
-  std::lock_guard<std::mutex> lock{mutex_};
-  return this->version_;
 }
 
 // implementation of threaded engine
