+++ tests/cpp/engine/threaded_engine_test.cc
@@ -275,64 +275,6 @@ TEST(Engine, basics) {
   LOG(INFO) << "All pass";
 }
 
-TEST(Engine, VarVersion) {
-  const size_t num_engines = 3;
-  std::vector<mxnet::Engine*> engines(num_engines);
-  engines[0] = mxnet::engine::CreateNaiveEngine();
-  engines[1] = mxnet::engine::CreateThreadedEnginePooled();
-  engines[2] = mxnet::engine::CreateThreadedEnginePerDevice();
-  std::string type_names[3] = {"NaiveEngine", "ThreadedEnginePooled", "ThreadedEnginePerDevice"};
-  for (size_t k = 0; k < num_engines; ++k) {
-    auto engine = engines[k];
-    std::vector<mxnet::Engine::OprHandle> oprs;
-
-    LOG(INFO) << "Testing var as a read dependency in " << type_names[k];
-    auto var = engine->NewVariable();
-    EXPECT_EQ(var->version(), 0U);
-    for (int i = 0; i < 10; ++i) {
-      oprs.push_back(engine->NewOperator(
-          [i](mxnet::RunContext ctx, mxnet::Engine::CallbackOnComplete cb) {
-            Foo(ctx, i);
-            cb();
-          },
-          {var}, {}));
-      engine->Push(oprs.at(i), mxnet::Context{});
-    }
-    engine->WaitForAll();
-    EXPECT_EQ(var->version(), 0U);
-    for (auto&& i : oprs) {
-      engine->DeleteOperator(i);
-    }
-    engine->DeleteVariable([](mxnet::RunContext) {}, mxnet::Context{}, var);
-    engine->WaitForAll();
-
-    LOG(INFO) << "Testing var as a write dependency in " << type_names[k];
-    var = engine->NewVariable();
-    EXPECT_EQ(var->version(), 0U);
-    oprs.clear();
-    for (int i = 0; i < 10; ++i) {
-      oprs.push_back(engine->NewOperator(
-          [i](mxnet::RunContext ctx, mxnet::Engine::CallbackOnComplete cb) {
-            Foo(ctx, i);
-            cb();
-          },
-          {}, {var}));
-      engine->Push(oprs.at(i), mxnet::Context{});
-    }
-    engine->WaitForAll();
-    EXPECT_EQ(var->version(), 10U);
-    for (auto&& i : oprs) {
-      engine->DeleteOperator(i);
-    }
-    engine->DeleteVariable([](mxnet::RunContext) {}, mxnet::Context{}, var);
-    engine->WaitForAll();
-
-    var = nullptr;
-    oprs.clear();
-    LOG(INFO) << "All pass";
-  }
-}
-
 #ifdef _OPENMP
 
 struct TestSaveAndRestoreOMPState {
