+++ tests/python/gpu/test_operator_gpu.py
@@ -41,7 +41,6 @@ from test_exc_handling import *
 from test_sparse_ndarray import *
 from test_sparse_operator import *
 from test_ndarray import *
-from test_subgraph_op import *
 
 set_default_context(mx.gpu(0))
 del test_support_vector_machine_l1_svm  # noqa
